[HER-16] Follow-up Rename user to body
