+++ src/admin/body/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { TableBody } from '@material-ui/core'
 
-import { users } from 'admin/user/users'
+import { users } from 'admin/body/users'
 import { Row } from 'admin/row'
 
 export const Body = () => <TableBody>
